Make code more go-lang friendly
